+++ src-tauri/src/lib.rs
@@ -11,7 +11,6 @@ use commands::WorkerState;
 
 pub fn run() {
     tauri::Builder::default()
-        .plugin(tauri_plugin_shell::init())
         .plugin(tauri_plugin_dialog::init())
         .invoke_handler(tauri::generate_handler![
             commands::get_worker_endpoint,